requirements.txt
flask==2.2.2
gunicorn==20.1.0
rich==12.5.1
requests==2.27.1
loguru==0.6.0
python-dotenv==0.19.2